scripts/requirements.txt
lxml==5.2.0
beautifulsoup4==4.12.0
numpy==1.26.4
pandas==2.1.0
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
pandas_datareader==0.10.0
